refactor!: Moved classes from 'core.expression' to 'core' package
The PlaceholderFinder and Matcher classes which were previously located in the 'core.expression' package have been moved to the 'core' package. The 'core.expression' package has subsequently been deleted. This refactoring effort was undertaken to improve the organization of the code.
