+++ engine/src/main/java/pro/verron/officestamper/core/Matcher.java
@@ -1,4 +1,4 @@
-package pro.verron.officestamper.core.expression;
+package pro.verron.officestamper.core;
 
 
 /**
