--- engine/src/main/java/pro/verron/officestamper/core/expression/package-info.java
@@ -1,4 +1,0 @@
-@NonNullApi
-package pro.verron.officestamper.core.expression;
-
-import org.springframework.lang.NonNullApi;
